requirements.txt
--index-url https://mirrors.aliyun.com/pypi/simple

modelscope==1.10
onnxruntime-gpu
faster_whisper
soundfile
kaldi_native_fbank
websockets
requests
pyyaml
jieba
scikit-learn
scipy